Move ClosingController to other class
